Change module structure
