+++ app/src/main/java/com/leeeyou/BaseFragment.kt
@@ -1,0 +1,13 @@
+package com.leeeyou
+
+import android.support.v4.app.Fragment
+
+/**
+ * @author:         leeeyou
+ * @description:    fragment base class
+ * @date:           2017/4/26
+ */
+abstract class BaseFragment : Fragment() {
+    abstract fun checkCanDoRefresh(): Boolean
+    abstract fun updateData()
+}
