+++ app/src/main/java/com/leeeyou/zhihudaily/model/bean/ZhiHuDailyItem.kt
@@ -1,0 +1,9 @@
+package com.leeeyou.zhihudaily.model.bean
+
+data class ZhiHuDailyItem(var images: List<String>,
+                          var type: Int,
+                          var id: Int,
+                          var ga_prefix: String,
+                          var title: String,
+                          var date: String
+)
