--- app/src/main/java/com/xyz/leeeyou/zhihuribao/data/service/ServiceFactory.kt
@@ -1,35 +1,0 @@
-package com.xyz.leeeyou.zhihuribao.data.service
-
-import retrofit.GsonConverterFactory
-import retrofit.Retrofit
-import retrofit.RxJavaCallAdapterFactory
-
-/**
- * ClassName: ServiceFactory
- * Description: The factory class mainly offers two kinds of Retrofit services, Rx mode and normal mode
- * 
- * Author:      leeeyou                             
- * Date:        2018/4/24 15:11
- */
-class ServiceFactory {
-    companion object {
-        fun <T> createRxRetrofitService(clazz: Class<T>, endPoint: String): T {
-            return Retrofit
-                    .Builder()
-                    .addCallAdapterFactory(RxJavaCallAdapterFactory.create())
-                    .addConverterFactory(GsonConverterFactory.create())
-                    .baseUrl(endPoint)
-                    .build()
-                    .create(clazz)
-        }
-
-        fun <T> createRetrofitService(clazz: Class<T>, endPoint: String): T {
-            return Retrofit
-                    .Builder()
-                    .addConverterFactory(GsonConverterFactory.create())
-                    .baseUrl(endPoint)
-                    .build()
-                    .create(clazz)
-        }
-    }
-}
