--- app/src/main/java/com/xyz/leeeyou/zhihuribao/view/activity/ZhiHuDailyDetailActivity.java
@@ -1,77 +1,0 @@
-package com.xyz.leeeyou.zhihuribao.view.activity;
-
-import android.os.Bundle;
-import android.webkit.WebView;
-
-import com.xyz.leeeyou.zhihuribao.R;
-import com.xyz.leeeyou.zhihuribao.data.model.ribao.ZhiHuDailyDetail;
-import com.xyz.leeeyou.zhihuribao.di.module.ZhiHuDailyRepositoryKt;
-import com.xyz.leeeyou.zhihuribao.utils.HtmlUtils;
-
-import rx.Subscriber;
-import rx.android.schedulers.AndroidSchedulers;
-import rx.functions.Action0;
-import rx.schedulers.Schedulers;
-
-public class ZhiHuDailyDetailActivity extends BaseOriginalActivity {
-
-    private int storyId;
-
-    WebView story_web;
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_story_detail);
-
-        story_web = findViewById(R.id.story_web);
-
-        storyId = getIntent().getIntExtra("storyId", -1);
-        String storyTitle = getIntent().getStringExtra("storyTitle");
-
-        setLeftTitleAndDisplayHomeAsUp(storyTitle);
-        initWebView();
-        getStoryDetail();
-    }
-
-    private void initWebView() {
-        story_web.setVerticalScrollBarEnabled(false);
-        story_web.getSettings().setDefaultTextEncodingName("UTF-8");
-    }
-
-
-    public void getStoryDetail() {
-        ZhiHuDailyRepositoryKt
-                .fetchStoryDetailById(storyId)
-                .subscribeOn(Schedulers.newThread())
-                .doOnSubscribe(new Action0() {
-                    @Override
-                    public void call() {
-//                        mMaterialDialog = new MaterialDialog.Builder(StoryDetailActivity.this)
-//                                .content("Loading ...")
-//                                .progress(true, 0)
-//                                .show();
-                    }
-                })
-                .subscribeOn(AndroidSchedulers.mainThread())
-                .observeOn(AndroidSchedulers.mainThread())
-                .subscribe(new Subscriber<ZhiHuDailyDetail>() {
-                    @Override
-                    public void onCompleted() {
-//                        mMaterialDialog.dismiss();
-                    }
-
-                    @Override
-                    public void onError(Throwable e) {
-//                        mMaterialDialog.dismiss();
-                        e.printStackTrace();
-                    }
-
-                    @Override
-                    public void onNext(ZhiHuDailyDetail storyDetail) {
-                        story_web.loadData(HtmlUtils.structHtml(storyDetail.getBody(), storyDetail.getCss()), "text/html; charset=UTF-8", null);
-                    }
-                });
-    }
-
-}
